requirements.txt
# Core dependencies
pref-voting
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.0  # Changed to be compatible with motor 3.6.0
pydantic==2.10.3
python-multipart==0.0.18

# Email service
postmarker==1.0  # For Postmark API

# Utilities
python-dotenv==1.0.1
arrow==1.3.0
humanize==4.11.0
func-timeout==4.3.5

# QR Code generation
qrcode[pil]==8.0
pillow==11.0.0

# Security
certifi==2024.8.30

# Math and algorithms (for voting)
nashpy==0.0.40
numba==0.60.0  # Works fine with Python 3.11